operation_tools/pyproject.toml
[tool.poetry]
name = "operation_tools"
version = "0.1.0"
description = "Operation tools for wam-v"
authors = ["Masaya Kataoka <ms.kataoka@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.10"
awscli-local = "^0.22.0"
awscli = "^1.34.5"
aws-cdk-lib = "^2.154.1"
constructs = "^10.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
